BugFix for updating screening
An error occured when updation screenings when selecting a time for a screening. E.g a 2 hour film required nothing on for 2 hours before even though the earlier film. Another bug that was fixed was the seats remaining for a screening when updating. IF a screening had no bookings then a null value was returned. Now changed to return a 0